fix(build+icons): Carbon Icons have name collisions
The reason for this is that the icons are placed into folders (16, 20, 24, 32) for sizes and some have the same name. We will use the folder name as part of the name to avoid collisions. This may be a breaking change for some, but Carbon Icons is relatively new, so hopefully doesn't disrupt many.