feat: use the Queue system for processing icon-sets
The previous version would run each of the icon-set processing in a new thread. But, there are many more icon-sets than cores on a computer. If we use a Queue system with `(cpus().length - 1 || 1)` for concurrency, then it will not peg out the computer and should be faster.
Also, we can now get total icons, total processing time, as well as generate an icon-info.json file.